4.3 debug

+++ pintos/src/filesys/filesys.c
@@ -62,7 +62,7 @@ filesys_create (const char *name, off_t initial_size)
   struct dir *dir = parse_path(cp_name, file_name);
   bool success = (dir != NULL
                   && free_map_allocate (1, &inode_sector)
-                  && inode_create (inode_sector, initial_size, (uint32_t)0)
+                  && inode_create (inode_sector, initial_size, 0)
                   && dir_add (dir, file_name, inode_sector));
   if (!success && inode_sector != 0) 
     free_map_release (inode_sector, 1);
